send email subscribers list
