+++ booking/main/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class MainConfig(AppConfig):
     name = 'main'
+
+    def ready(self):
+        import main.signals
