+++ booking/main/templates/main/mail_subscribers.html
@@ -1,0 +1,4 @@
+<h3>Зарегестрирован новый объект размещения</h3>
+<p>Отель {{ room.hotel.title }}, комната № {{ room.number }}, площадь - {{ room.area }}</p>
+<p>Услуги в номере: {{ room.tags.tag_name }}</p>
+<p>Описание - <a fref="#">ссылка</a></p>
